Drop support for ruby 1.9.3
According to the
[Rails Hosting Survey](http://rails-hosting.com/2016/index.html),
usage of ruby 1.9.3 has dropped to around 10%, with much of
this drop occuring in 2016.

> Usage of 1.9.3 dropped by 16% and adoption of Ruby 2.2 and 2.3 jumped
> up to 27% each.

The 2016 survey had 1,417 respondents.

IMO, it is acceptable for development of a new major version of
authlogic, version 4, to continue, while the 10% of the community
using 1.9.3 continues to use authlogic 3.

Semver does not specify how to treat this change, and there is no consensus
that I'm aware of in the ruby community. I recommend a major version bump.
Version numbers are free, after all.