Add Authlogic::Session::Cookies#protect_from_replay config option
This will encrypt the user_credentials cookie issued to a user using AES with the session_id as key.

If the remember_me option is activated, the user_credentials won't be encrypted because remember_me needs the user_credentials cookie to stay valid even after the session_id is changed.