Introduce a debugger, byebug
I don't care which debugger we use. If y'all want to use pry, and it works
(I've had issues) that's fine. Also, if there's a *convenient* way to use a
debugger without putting it in the gemspec, let me know.

The version constraint is ~> 9.0 because we still support ruby 2.1.
When we drop support for ruby 2.1, we can constrain byebug to ~> 9.1.