Backport: Deprecate authenticates_many
In early 2017, when we discussed our "wishlists", Ben mentioned that
he would like to drop `authenticates_many`. At that time, I was using
it at work, but I am not using it any longer.

It's a useful feature, but I'd like to see Authlogic "core" do less.

As the message says, if someone wants to extract and maintain a
new gem, that'd be fine. I'd help someone get started with that.